Changed from degrees to radians in the renderer

+++ Volum/src/Volum/Renderer/Renderer2D.cpp
@@ -129,7 +129,7 @@ namespace Volum
 	{
 		VLM_PROFILE_FUNCTION();
 
-		uint32_t dataSize = (uint8_t*)s_data.QuadVertexBufferPtr - (uint8_t*)s_data.QuadVertexBufferBase; // cast to 1 byte type to get size in bytes
+		uint32_t dataSize = (uint32_t)((uint8_t*)s_data.QuadVertexBufferPtr - (uint8_t*)s_data.QuadVertexBufferBase); // cast to 1 byte type to get size in bytes
 		s_data.QuadVertexBuffer->SetData(s_data.QuadVertexBufferBase, dataSize);
 
 		for (uint32_t i = 0; i < s_data.TextureSlotIndex; i++)
@@ -285,7 +285,7 @@ namespace Volum
 		const float textureTilingFactor = 1.0f;
 
 		glm::mat4 transform = glm::translate(glm::mat4(1.0f), position)
-			* glm::rotate(glm::mat4(1.0f), glm::radians(rotation), { 0.0f, 0.0f, 1.0f })
+			* glm::rotate(glm::mat4(1.0f), rotation, { 0.0f, 0.0f, 1.0f })
 			* glm::scale(glm::mat4(1.0f), { size.x, size.y, 1.0f });
 
 		s_data.QuadVertexBufferPtr->Position = transform * s_data.QuadVertexPositions[0];
@@ -351,7 +351,7 @@ namespace Volum
 		}
 
 		glm::mat4 transform = glm::translate(glm::mat4(1.0f), position)
-			* glm::rotate(glm::mat4(1.0f), glm::radians(rotation), glm::vec3(0.0f, 0.0f, 1.0f))
+			* glm::rotate(glm::mat4(1.0f), rotation, glm::vec3(0.0f, 0.0f, 1.0f))
 			* glm::scale(glm::mat4(1.0f), { size.x, size.y, 1.0f });
 
 		s_data.QuadVertexBufferPtr->Position = transform * s_data.QuadVertexPositions[0];
